Drt 4568 egate csv percentage

+++ client/src/main/scala/spatutorial/client/components/TerminalPage.scala
@@ -14,6 +14,7 @@ import japgolly.scalajs.react.extra.router.RouterCtl
 import japgolly.scalajs.react.vdom.html_<^._
 import japgolly.scalajs.react.vdom.{TagOf, VdomArray, html_<^}
 import org.scalajs.dom.html.Div
+import drt.client.components.FlightTableComponents
 
 import scala.util.Try
 
@@ -115,7 +116,7 @@ TerminalPage {
 object FlightComponents {
 
   def paxComp(maxFlightPax: Int = 853)(flight: ApiFlight, apiSplits: ApiSplits): TagMod = {
-    val apiPax: Int = apiSplits.splits.map(_.paxCount).sum
+    val apiPax: Int = apiSplits.splits.map(_.paxCount.toInt).sum
 
     val (paxNos, paxClass, paxWidth) = if (apiPax > 0)
       (apiPax, "pax-api", paxBarWidth(maxFlightPax, apiPax))
